tests/wpt/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/stroke-dasharray.html

<!doctype html>
<meta charset="utf-8">
<title>'stroke-dasharray' property</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-get">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-set">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#property-stle-value-normalization">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<script src="resources/testsuite.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

runPropertyTests('stroke-dasharray', [
  { syntax: 'none' },
  {
    syntax: '<length>',
    specified: assert_is_equal_with_range_handling,
  },
  {
    syntax: '<percentage>',
    specified: assert_is_equal_with_range_handling,
  },
  {
    syntax: '<number>',
    specified: assert_is_equal_with_range_handling,
    computed: (_, result) => assert_is_unit('px', result)
  },
]);

runUnsupportedPropertyTests('stroke-dasharray', [
  '5% 1em 2%',
]);

</script>
